add splashscreen as a component and fade out effect
