+++ frontend/src/styles/pages/Home.module.scss
@@ -2,6 +2,19 @@
 @use '@styles/settings/media' as media;
 @use '@styles/tools/focus';
 
+.splashScreen {
+    position: fixed;
+    inset: 0;
+    opacity: 1;
+    visibility: visible;
+    transition: opacity 0.6s ease-out, visibility 1s ease-out;
+}
+
+.hide {
+    opacity: 0;
+    visibility: hidden;
+}
+
 .container {
     @include container.container;
 }
