--- frontend/src/styles/pages/SplashScreen.module.scss
@@ -1,40 +1,0 @@
-.splashScreen {
-  background-image: url("../../../src/assets/images/splash-screen-background.svg");
-  background-repeat: no-repeat;
-  background-size: cover;
-  background-position: 50% 50%;
-  min-height: 100vh;
-
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-
-.splashScreen h1 {
-  font-size: 10rem;
-  font-family: var(--font-title);
-  transform: translateX(90px);
-}
-
-.splashScreen span {
-  display: inline-block;
-}
-
-.heading1 {
-  color: var(--colour-greenBlack);
-  visibility: hidden;
-}
-
-.heading2 {
-  color: var(--colour-darkGreen);
-  visibility: hidden;
-}
-
-.imgcontainer {
-  visibility: hidden;
-  transform: translate(-17.5px, -10px);
-}
-
-.logo {
-  min-height: 310px;
-}
